refactor: extract components and move into folder
