+++ src/components/Game/Board/Markers/EmptyMarker.tsx
@@ -1,0 +1,7 @@
+const EmptyMarker = () => {
+  return (
+    <div className="w-[70px] h-[75px] tablet:w-[70px] tablet:h-[75px] mobile:w-[41px] mobile:h-[46px] mini:w-[42px] mini:h-[46px]" />
+  );
+};
+
+export default EmptyMarker;
